+++ twilio/rest/insights/v1/call/metric.py
@@ -37,15 +37,15 @@ class MetricInstance(InstanceResource):
         CLIENT_EDGE = "client_edge"
 
     """
-    :ivar timestamp: 
-    :ivar call_sid: 
-    :ivar account_sid: 
+    :ivar timestamp: Timestamp of metric sample. Samples are taken every 10 seconds and contain the metrics for the previous 10 seconds.
+    :ivar call_sid: The unique SID identifier of the Call.
+    :ivar account_sid: The unique SID identifier of the Account.
     :ivar edge: 
     :ivar direction: 
-    :ivar carrier_edge: 
-    :ivar sip_edge: 
-    :ivar sdk_edge: 
-    :ivar client_edge: 
+    :ivar carrier_edge: Contains metrics and properties for the Twilio media gateway of a PSTN call.
+    :ivar sip_edge: Contains metrics and properties for the Twilio media gateway of a SIP Interface or Trunking call.
+    :ivar sdk_edge: Contains metrics and properties for the SDK sensor library for Client calls.
+    :ivar client_edge: Contains metrics and properties for the Twilio media gateway of a Client call.
     """
 
     def __init__(self, version: Version, payload: Dict[str, Any], call_sid: str):
@@ -103,7 +103,7 @@ class MetricList(ListResource):
         Initialize the MetricList
 
         :param version: Version that contains the resource
-        :param call_sid:
+        :param call_sid: The unique SID identifier of the Call.
 
         """
         super().__init__(version)
@@ -127,8 +127,8 @@ class MetricList(ListResource):
         is reached.
         The results are returned as a generator, so this operation is memory efficient.
 
-        :param &quot;MetricInstance.TwilioEdge&quot; edge:
-        :param &quot;MetricInstance.StreamDirection&quot; direction:
+        :param &quot;MetricInstance.TwilioEdge&quot; edge: The Edge of this Metric. One of `unknown_edge`, `carrier_edge`, `sip_edge`, `sdk_edge` or `client_edge`.
+        :param &quot;MetricInstance.StreamDirection&quot; direction: The Direction of this Metric. One of `unknown`, `inbound`, `outbound` or `both`.
         :param limit: Upper limit for the number of records to return. stream()
                       guarantees to never return more than limit.  Default is no limit
         :param page_size: Number of records to fetch per request, when not set will use
@@ -156,8 +156,8 @@ class MetricList(ListResource):
         is reached.
         The results are returned as a generator, so this operation is memory efficient.
 
-        :param &quot;MetricInstance.TwilioEdge&quot; edge:
-        :param &quot;MetricInstance.StreamDirection&quot; direction:
+        :param &quot;MetricInstance.TwilioEdge&quot; edge: The Edge of this Metric. One of `unknown_edge`, `carrier_edge`, `sip_edge`, `sdk_edge` or `client_edge`.
+        :param &quot;MetricInstance.StreamDirection&quot; direction: The Direction of this Metric. One of `unknown`, `inbound`, `outbound` or `both`.
         :param limit: Upper limit for the number of records to return. stream()
                       guarantees to never return more than limit.  Default is no limit
         :param page_size: Number of records to fetch per request, when not set will use
@@ -186,8 +186,8 @@ class MetricList(ListResource):
         Unlike stream(), this operation is eager and will load `limit` records into
         memory before returning.
 
-        :param &quot;MetricInstance.TwilioEdge&quot; edge:
-        :param &quot;MetricInstance.StreamDirection&quot; direction:
+        :param &quot;MetricInstance.TwilioEdge&quot; edge: The Edge of this Metric. One of `unknown_edge`, `carrier_edge`, `sip_edge`, `sdk_edge` or `client_edge`.
+        :param &quot;MetricInstance.StreamDirection&quot; direction: The Direction of this Metric. One of `unknown`, `inbound`, `outbound` or `both`.
         :param limit: Upper limit for the number of records to return. list() guarantees
                       never to return more than limit.  Default is no limit
         :param page_size: Number of records to fetch per request, when not set will use
@@ -218,8 +218,8 @@ class MetricList(ListResource):
         Unlike stream(), this operation is eager and will load `limit` records into
         memory before returning.
 
-        :param &quot;MetricInstance.TwilioEdge&quot; edge:
-        :param &quot;MetricInstance.StreamDirection&quot; direction:
+        :param &quot;MetricInstance.TwilioEdge&quot; edge: The Edge of this Metric. One of `unknown_edge`, `carrier_edge`, `sip_edge`, `sdk_edge` or `client_edge`.
+        :param &quot;MetricInstance.StreamDirection&quot; direction: The Direction of this Metric. One of `unknown`, `inbound`, `outbound` or `both`.
         :param limit: Upper limit for the number of records to return. list() guarantees
                       never to return more than limit.  Default is no limit
         :param page_size: Number of records to fetch per request, when not set will use
@@ -251,8 +251,8 @@ class MetricList(ListResource):
         Retrieve a single page of MetricInstance records from the API.
         Request is executed immediately
 
-        :param edge:
-        :param direction:
+        :param edge: The Edge of this Metric. One of `unknown_edge`, `carrier_edge`, `sip_edge`, `sdk_edge` or `client_edge`.
+        :param direction: The Direction of this Metric. One of `unknown`, `inbound`, `outbound` or `both`.
         :param page_token: PageToken provided by the API
         :param page_number: Page Number, this value is simply for client state
         :param page_size: Number of records to return, defaults to 50
@@ -284,8 +284,8 @@ class MetricList(ListResource):
         Asynchronously retrieve a single page of MetricInstance records from the API.
         Request is executed immediately
 
-        :param edge:
-        :param direction:
+        :param edge: The Edge of this Metric. One of `unknown_edge`, `carrier_edge`, `sip_edge`, `sdk_edge` or `client_edge`.
+        :param direction: The Direction of this Metric. One of `unknown`, `inbound`, `outbound` or `both`.
         :param page_token: PageToken provided by the API
         :param page_number: Page Number, this value is simply for client state
         :param page_size: Number of records to return, defaults to 50