+++ twilio/rest/insights/v1/setting.py
@@ -24,10 +24,10 @@ from twilio.base.version import Version
 class SettingInstance(InstanceResource):
 
     """
-    :ivar account_sid:
-    :ivar advanced_features:
-    :ivar voice_trace:
-    :ivar url:
+    :ivar account_sid: The unique SID identifier of the Account.
+    :ivar advanced_features: A boolean flag indicating whether Advanced Features for Voice Insights are enabled.
+    :ivar voice_trace: A boolean flag indicating whether Voice Trace is enabled.
+    :ivar url: The URL of this resource.
     """
 
     def __init__(self, version: Version, payload: Dict[str, Any]):
@@ -60,7 +60,7 @@ class SettingInstance(InstanceResource):
         """
         Fetch the SettingInstance
 
-        :param subaccount_sid:
+        :param subaccount_sid: The unique SID identifier of the Subaccount.
 
         :returns: The fetched SettingInstance
         """
@@ -74,7 +74,7 @@ class SettingInstance(InstanceResource):
         """
         Asynchronous coroutine to fetch the SettingInstance
 
-        :param subaccount_sid:
+        :param subaccount_sid: The unique SID identifier of the Subaccount.
 
         :returns: The fetched SettingInstance
         """
@@ -91,9 +91,9 @@ class SettingInstance(InstanceResource):
         """
         Update the SettingInstance
 
-        :param advanced_features:
-        :param voice_trace:
-        :param subaccount_sid:
+        :param advanced_features: A boolean flag to enable Advanced Features for Voice Insights.
+        :param voice_trace: A boolean flag to enable Voice Trace.
+        :param subaccount_sid: The unique SID identifier of the Subaccount.
 
         :returns: The updated SettingInstance
         """
@@ -112,9 +112,9 @@ class SettingInstance(InstanceResource):
         """
         Asynchronous coroutine to update the SettingInstance
 
-        :param advanced_features:
-        :param voice_trace:
-        :param subaccount_sid:
+        :param advanced_features: A boolean flag to enable Advanced Features for Voice Insights.
+        :param voice_trace: A boolean flag to enable Voice Trace.
+        :param subaccount_sid: The unique SID identifier of the Subaccount.
 
         :returns: The updated SettingInstance
         """
@@ -151,7 +151,7 @@ class SettingContext(InstanceContext):
         """
         Fetch the SettingInstance
 
-        :param subaccount_sid:
+        :param subaccount_sid: The unique SID identifier of the Subaccount.
 
         :returns: The fetched SettingInstance
         """
@@ -175,7 +175,7 @@ class SettingContext(InstanceContext):
         """
         Asynchronous coroutine to fetch the SettingInstance
 
-        :param subaccount_sid:
+        :param subaccount_sid: The unique SID identifier of the Subaccount.
 
         :returns: The fetched SettingInstance
         """
@@ -204,9 +204,9 @@ class SettingContext(InstanceContext):
         """
         Update the SettingInstance
 
-        :param advanced_features:
-        :param voice_trace:
-        :param subaccount_sid:
+        :param advanced_features: A boolean flag to enable Advanced Features for Voice Insights.
+        :param voice_trace: A boolean flag to enable Voice Trace.
+        :param subaccount_sid: The unique SID identifier of the Subaccount.
 
         :returns: The updated SettingInstance
         """
@@ -235,9 +235,9 @@ class SettingContext(InstanceContext):
         """
         Asynchronous coroutine to update the SettingInstance
 
-        :param advanced_features:
-        :param voice_trace:
-        :param subaccount_sid:
+        :param advanced_features: A boolean flag to enable Advanced Features for Voice Insights.
+        :param voice_trace: A boolean flag to enable Voice Trace.
+        :param subaccount_sid: The unique SID identifier of the Subaccount.
 
         :returns: The updated SettingInstance
         """